Cambiar valor a floato del precio de los mangas

+++ grails-app/domain/Colecciones/Mangas.groovy
@@ -14,7 +14,7 @@ class Mangas {
 	
 	int numTomosActuales
 	int numTomosMaximos
-	int precioTotal
+	float precioTotal
 
 	Date fechaInscripcion
 	Date ultimaModificacion
